Safari-609.1.11.1/LayoutTests/http/tests/eventsource/eventsource-cors-no-server-expected.txt

Test that EventSource tries to reconnect if there's no server response when making cross-origin requests. Should print a series of PASS messages followed by DONE.

PASS: got error event and readyState is CONNECTING
